Don't try to parse ticket for all actions
I initially hardcoded ticket parsing while parsing the response to
an action request since this was easier to quickly test the code.
However this is limited and wrong. This commit adds an optional
function pointer parameter to the response parsing code so that
action-specific behaviour can be done while parsing the response.