Handle user/admin mode in REST calls
This is controlled through a 'Filter' header, it should be set to
true when accessing the API as a user, and to false when accessing
it as an admin. Access to the admin API will be refused if the user
you are logging in as is not an admin.
This header is automatically added to the REST calls depending on
the value of the OvirtProxy::admin property.