OvirtResourceRestCall: Handle RestParams
If there are some RestParams set on tho OvirtResourceRestCall, we
handle them by adding them in the query part of the URI. This will
be used when updating 'cdroms' (subcollections of vms) as they
can use an optional 'current' parameter.
This requires a patch to librest though so that it can handle
rest_proxy_call_set_function being called from a
RestProxyCall::serialize_params vfunc.