Don't force ssl-strict to FALSE on RestProxy instances
Setting this to FALSE causes most checks on the remote host SSL
certificates to be disabled (in particular whether it should be checked
against trusted root certificates or not). Since this is not wanted
in most cases, and this can open the door to man-in-the-middle attacks,
it's better to not set it to FALSE, especially now that we can pass
the CA certs to use down to librest/libsoup.
OvirtProxy derives from RestProxy, so client applications can disable
this themselves if they wish to.