OvirtProxy: take into account RestProxy::ssl-ca-cert
Now that RestProxy has a ssl-ca-cert property, OvirtProxy::ca-cert
is redundant.  We want to use ssl-ca-file as the canonical property, and
set ca-cert value from it. However, when the user sets the ca-cert
property, we need to create a temporary file in order to be able to keep
the ssl-ca-file property synced with it, which is why the code changes
are more complicated than expected.