Fix setting of OvirtProxy::ssl-ca-file from tmp file
We were setting the path for the temporary file using g_object_set,
which triggered a callback listening for ssl-ca-file changes, which
made sure the temp file was removed when ssl-ca-file is modified.
We obviously do not want to do that when we are currently setting
the path to the temporary file we just created, so we just block
this handler while setting it.