Fix memory corruption when RestProxy::ssl-ca-file does not exist
Older librest releases don't have the RestProxy::ssl-ca-file property.
Without this property, we are not going to be able to set the CA
certificate to use to validate the oVirt instance TLS certificate, but we
should at least be able to report a connection error.
However, g_object_get() will not set its parameter to NULL when the
property does not exist, so we need to set it to NULL ourselves before
calling g_object_get() if we want to rely on that.

If we don't do that, 'ca_file' will be uninitialized when
RestProxy::ssl-ca-file does not exist, and get_ca_cert_data() will call
g_free() on that variable, causing memory corruption.