Sanitize append of '/api' to oVirt URI
"/api" should not be part of the base oVirt REST URL, but is the
REST function which is used to access the OvirtApi entry point. The
current code expects an URI ending in /api, and then strips it from the
URI in various places.
This commit cleans this up by making sure the base URI does not include
"/api" and by adding it in appropriate places. It also automatically
appends https:// to the give host.