Fix crash on failure to parse response to an action
When a REST action fails with a 404 HTTP error (for example), the
document we get back is not an XML document. Trying to parse it will
return a NULL root node. If we don't return early in such a situation,
we'll get a crash trying to dereference it later on.