Add OvirtResource::add_rest_params
This virtual method (wrapped by ovirt_resource_add_rest_params) is used
by OvirtResourceRestCall to let classes deriving from OvirtResource add
some class-specific arguments to the OvirtResourceRestCall that is going
to be run.

This will be used by OvirtCdrom to append a ";current" to the request it
makes in order to only manage the current cdrom image, not the one which
was set at boot time.