Add ovirt_cdrom_add_rest_params()
At the moment this unconditionnally set ;current on the
OvirtResourceRestCalls made on an OvirtCdrom resource. Ideally
the library user would be able to choose whether to manage the current
or boot-time cdrom image. When this is possible, ovirt_cdrom_update()
will be obsolete and can be replaced by the generic
ovirt_resource_update().