Fix return value of ovirt_rest_call_finish()
In some cases (when the caller did not provide an additional callback
to be called on completion of the async operation),
ovirt_rest_call_finish() would return FALSE even though the async call
was successful. This could be hit when using the async version of
ovirt_cdrom_update_async() for example.