storage-domain: Silence warning during XML parsing
Currently, the storage domain XML parser emits a warning when it can't
find one of the elements it's looking for. However, all of these are
marked as optional in the corresponding XML schema, so printing a
runtime warning about that is wrong. This commit switches it from
g_warning() to g_debug().