Add GOVIRT_NO_SSL_STRICT debug environment variable
During development or for debugging, it might be insteresting to access
the oVirt REST API without a full TLS setup (ie with invalid
certificates) as this can be a pain to put in place. The
GOVIRT_NO_SSL_STRICT environment variable instructs libgovirt not to
fail when it encounters invalid certificates during an HTTPS session.