proxy: Update display CA in fetch_certificate{_async,}
There is an historical confusion between the CA used for REST
communications, and the CA used for SPICE/display communication.
https://ovirt.example.com/ca.crt is the CA certificate used for SPICE
communication, and may be the one used for REST communication.
Now that we have distinct CA certs on OvirtVmDisplay and OvirtProxy, we
need to make sure that ovirt_proxy_fetch_certificate{_async,} which
fetches ca.crt is properly updating OvirtVmDisplay::ca-cert as this is
the one which was fetched.