proxy: Add ovirt_proxy_get_vms_internal()
This internal function will be used by several of the subsequent commits
in order to get the list of VMs after going through
OvirtProxy->OvirtApi->OvirtCollection->VM GList

The _internal() name is just a hack because the same
ovirt_proxy_get_vms() is already exported but is deprecated, so using it
would cause warnings. We could also do some #ifdef magic to make sure
ovirt_proxy_get_vms() do not appear deprecated when libgovirt is being
compiled, but for now this _internal() suffix will do.