proxy: Improve NULL CA handling in set_tmp_ca_file
We make sure not to set RestProxy::ssl-ca-file when the 'ca_file'
argument to ovirt_proxy_set_tmp_ca_file() is NULL. This was added in
order to prevent an infinite notification loop involving
ssl_ca_file_changed(). However, since the previous commit, this infinite
loop no longer happens, and not clearing RestProxy::ssl-ca-file when we
have removed the corresponding temporary file results in an inconsistent
situation.