proxy: Don't set RestProxy::ssl-ca-file from ssl_ca_file_changed()
The only purpose of our notify::ssl-ca-file handler is to cleanup the
temporary file we created when the ssl-ca-file is set externally. Now
that we have a separate ovirt_proxy_free_tmp_ca_file(), we can call that
rather than relying on ovirt_proxy_set_tmp_ca_file(proxy, NULL); to
achieve that.