proxy: Fix bug in 'cancelled' disconnection after async calls
9aa3aec8 added disconnection of the 'cancelled' signal on the
cancellable in async calls once the async call has completed to prevent
a crash. However, there was a bug in this patch as we only disconnect
from 'cancelled' if data->cancellable is not NULL, and there was one
code path where setting data->cancellable was forgotten.
A few refactorings later, the buggy codepath is the only one which
remains, so the bug fixed in 9aa3aec8 is back...

This commit makes sure data->cancellable is set when needed.