Revert "proxy: Improve NULL CA handling in set_tmp_ca_file"
This reverts commit 5c8f3c33e6ceb46d14a501dae9f03c40eb81ef49.

Being able to set a NULL CA certificate after setting a non-NULL one
would require some changes in libsoup. Since we are using a deprecated
libsoup property (ssl-ca-file), this is unlikely to go upstream at this
point, see https://bugzilla.gnome.org/show_bug.cgi?id=754825
As this also causes breakage with upstream libsoup (certificate check
failures when trying to connect to an oVirt instance), it's better to
revert this patch for now.