proxy: Check if operation is cancelled before disconnecting signal
According to the documentation, g_cancellable_disconnect() waits for the
signal handler to finish, and if it is called from the handler itself, it
will result in a deadlock. To avoid it, we check if the operation is
cancelled and if so, call g_signal_handler_disconnect() instead of
g_cancellable_disconnect().

https://developer.gnome.org/gio/stable/GCancellable.html#g-cancellable-disconnect

Signed-off-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>