utils: Retrieve node attributes in ovirt_resource_parse_xml()
This commit adds a new field to the OvirtXmlElement struct, which is
used to retrieve an attribute from the xml node. It is optional, meaning
that, if not informed, the function will still retrieve the node
contents instead.

We also introduce ovirt_rest_xml_node_get_attr_from_path() function,
to retrieve the given attribute value from the path below the node, much
similar to ovirt_rest_xml_node_get_content_from_path().

Signed-off-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>