Move resource type definitions to ovirt-types.h
With recently added ovirt_resource_new* functions, it is now possible to
create any type of resource, being only necessary to pass the type as
argument.

In order to avoid interdependencies between objects, we need a common
place for definition of the resource types. For instance, it should be
possible to access a OvirtHost from a OvirtVm and vice-versa.

Signed-off-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>