proxy: Set detailed error message for async call
The rest API returns more detailed error messages with the result, not
only the literal corresponding to the value. If this is the case, we set
a new error message and return it.

For example, before this change, virt-viewer showed a dialog with a
vague 'Bad Request' message, and now a much more detailed 'Operation
Failed: query execution failed due to insufficient permissions.'

Signed-off-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>